requirements.txt
python-telegram-bot[job-queue]
pygsheets==2.0.6
python-dotenv==1.0.0